Core: fix IP_t size to make party join work correctly (#82)
On linux x64, sizeof unsigned long is 8
While on window, the client part, sizeof unsigned long is 4
This mismatch make the GCPartyJoined packet parsing error

+++ src/Core/types/SystemTypes.h
@@ -110,7 +110,7 @@ enum WorldStatus
 	WORLD_CLOSE	
 };
 
-typedef unsigned long IP_t;
+typedef unsigned int IP_t;
 const uint szIP = sizeof(IP_t);
 
 typedef BYTE WorldID_t;
